+++ src/position.rs
@@ -1,17 +1,10 @@
-use std::ops::{AddAssign, Mul};
-use super::{lsb, consts::*, movegen::{bishop_attacks, rook_attacks}, zobrist::ZVALS};
+use super::{lsb, consts::*, zobrist::ZVALS};
 
 macro_rules! from {($m:expr) => {(($m >> 6) & 63) as usize}}
 macro_rules! to {($m:expr) => {($m & 63) as usize}}
 macro_rules! bit {($x:expr) => {1 << $x}}
 macro_rules! pop {($x:expr) => {$x &= $x - 1}}
 
-/// Main position struct:
-/// - Holds all information needed for the board state
-/// - 6 piece bitboards and 2 colour bitboards
-/// - Mailbox array for finding pieces quickly
-/// - Incrementally updated zobrist hash, phase and endgame and midgame
-/// piece-square table scores
 pub struct Position {
     pub pieces: [u64; 6],
     pub sides: [u64; 2],
@@ -43,21 +36,38 @@ pub struct MoveContext {
     captured_pc: u8,
 }
 
-#[derive(Clone, Copy, Debug, Default)]
-pub struct S(pub i16, pub i16);
-
-impl AddAssign<S> for S {
-    fn add_assign(&mut self, rhs: S) {
-        self.0 += rhs.0;
-        self.1 += rhs.1;
-    }
+#[inline(always)]
+pub fn rook_attacks(idx: usize, occ: u64) -> u64 {
+    let m: Mask = RMASKS[idx];
+    let mut f: u64 = occ & m.file;
+    let mut r: u64 = f.swap_bytes();
+    f -= m.bit;
+    r -= m.bit.swap_bytes();
+    f ^= r.swap_bytes();
+    f &= m.file;
+    let mut e: u64 = m.right & occ;
+    r = e & e.wrapping_neg();
+    e = (r ^ (r - m.bit)) & m.right;
+    let w: u64 = m.left ^ WEST[(((m.left & occ)| 1).leading_zeros() ^ 63) as usize];
+    f | e | w
 }
 
-impl Mul<S> for i16 {
-    type Output = S;
-    fn mul(self, rhs: S) -> Self::Output {
-        S(self * rhs.0, self * rhs.1)
-    }
+#[inline(always)]
+pub fn bishop_attacks(idx: usize, occ: u64) -> u64 {
+    let m: Mask = BMASKS[idx];
+    let mut f: u64 = occ & m.right;
+    let mut r: u64 = f.swap_bytes();
+    f -= m.bit;
+    r -= m.file;
+    f ^= r.swap_bytes();
+    f &= m.right;
+    let mut f2: u64 = occ & m.left;
+    r = f2.swap_bytes();
+    f2 -= m.bit;
+    r -= m.file;
+    f2 ^= r.swap_bytes();
+    f2 &= m.left;
+    f | f2
 }
 
 impl Position {
@@ -242,10 +252,6 @@ impl Position {
         self.c = !self.c;
     }
 
-    pub fn fifty_draw(&self) -> bool {
-        self.state.halfmove_clock >= 100
-    }
-
     pub fn repetition_draw(&self, num: u8) -> bool {
         let l: usize = self.stack.len();
         if l < 6 || self.nulls > 0 { return false }
@@ -271,11 +277,5 @@ impl Position {
             return true
         }
         false
-    }
-
-    pub fn mvv_lva(&self, m: u16) -> u16 {
-        let moved_pc: usize = self.squares[from!(m)] as usize;
-        let captured_pc: usize = self.squares[to!(m)] as usize;
-        MVV_LVA[captured_pc][moved_pc]
     }
 }
